using shared model layer
