+++ Kanban.sln
@@ -19,19 +19,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Application", "src\K
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Repository", "src\Kanban.Repository\Kanban.Repository.csproj", "{8D0BE32F-02D3-4533-9755-C9D77726DB75}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.API.Dto", "src\API.Dto\Kanban.API.Dto.csproj", "{DC363BB3-7574-485C-91D2-ECC9F148C2FB}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Application.Dto", "src\Kanban.Application.Dto\Kanban.Application.Dto.csproj", "{65CCF79B-796D-4820-A352-48D8EFF2620F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Repository.Dto", "src\Kanban.Repository.Dto\Kanban.Repository.Dto.csproj", "{3F118892-E8F0-4DA8-9DC0-1BFF97320286}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1 - API", "1 - API", "{48793681-C683-4734-A5F7-8BA6AE27C09C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2 - Application", "2 - Application", "{13FB0C36-4889-4908-857D-2A0FCCB7D338}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "3 - Infrastructure", "3 - Infrastructure", "{69D7D72B-851F-4245-B68E-1BE2BC5F753D}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "4 - Tests", "4 - Tests", "{E98514FF-0523-49B4-8564-8D196AE6B640}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "6 - Tests", "6 - Tests", "{E98514FF-0523-49B4-8564-8D196AE6B640}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Application.Tests", "tests\Kanban.Application.Tests\Kanban.Application.Tests.csproj", "{431B93E3-0CE4-4BDD-9447-04E09667A39E}"
 EndProject
@@ -44,6 +38,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Integration.Tests", "tests\Kanban.Integration.Tests\Kanban.Integration.Tests.csproj", "{C41D0C1D-4400-408D-B45E-82251EF7E4A1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kanban.Presentation.Tests", "tests\Kanban.Presentation.Tests\Kanban.Presentation.Tests.csproj", "{292874FF-1ECC-4F19-BFB9-B047D26059C5}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "4 - Model", "4 - Model", "{9721A4D2-6BF8-4A73-B2D9-3C416D5A8458}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kanban.Model", "src\Kanban.Model\Kanban.Model.csproj", "{3A18D620-D1A6-4188-BF64-48CCF3A8E61D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -63,18 +61,6 @@ Global
 		{8D0BE32F-02D3-4533-9755-C9D77726DB75}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D0BE32F-02D3-4533-9755-C9D77726DB75}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D0BE32F-02D3-4533-9755-C9D77726DB75}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DC363BB3-7574-485C-91D2-ECC9F148C2FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DC363BB3-7574-485C-91D2-ECC9F148C2FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DC363BB3-7574-485C-91D2-ECC9F148C2FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DC363BB3-7574-485C-91D2-ECC9F148C2FB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{65CCF79B-796D-4820-A352-48D8EFF2620F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{65CCF79B-796D-4820-A352-48D8EFF2620F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{65CCF79B-796D-4820-A352-48D8EFF2620F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{65CCF79B-796D-4820-A352-48D8EFF2620F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F118892-E8F0-4DA8-9DC0-1BFF97320286}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F118892-E8F0-4DA8-9DC0-1BFF97320286}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F118892-E8F0-4DA8-9DC0-1BFF97320286}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F118892-E8F0-4DA8-9DC0-1BFF97320286}.Release|Any CPU.Build.0 = Release|Any CPU
 		{431B93E3-0CE4-4BDD-9447-04E09667A39E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{431B93E3-0CE4-4BDD-9447-04E09667A39E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{431B93E3-0CE4-4BDD-9447-04E09667A39E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -95,6 +81,10 @@ Global
 		{292874FF-1ECC-4F19-BFB9-B047D26059C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{292874FF-1ECC-4F19-BFB9-B047D26059C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{292874FF-1ECC-4F19-BFB9-B047D26059C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3A18D620-D1A6-4188-BF64-48CCF3A8E61D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A18D620-D1A6-4188-BF64-48CCF3A8E61D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A18D620-D1A6-4188-BF64-48CCF3A8E61D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3A18D620-D1A6-4188-BF64-48CCF3A8E61D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -103,14 +93,12 @@ Global
 		{7674C0A6-5822-44DB-BBE7-0191820FD198} = {48793681-C683-4734-A5F7-8BA6AE27C09C}
 		{1BAF0DCB-7039-4308-A055-232D3A732F79} = {13FB0C36-4889-4908-857D-2A0FCCB7D338}
 		{8D0BE32F-02D3-4533-9755-C9D77726DB75} = {69D7D72B-851F-4245-B68E-1BE2BC5F753D}
-		{DC363BB3-7574-485C-91D2-ECC9F148C2FB} = {48793681-C683-4734-A5F7-8BA6AE27C09C}
-		{65CCF79B-796D-4820-A352-48D8EFF2620F} = {13FB0C36-4889-4908-857D-2A0FCCB7D338}
-		{3F118892-E8F0-4DA8-9DC0-1BFF97320286} = {69D7D72B-851F-4245-B68E-1BE2BC5F753D}
 		{431B93E3-0CE4-4BDD-9447-04E09667A39E} = {E98514FF-0523-49B4-8564-8D196AE6B640}
 		{1D651461-839D-4648-85B9-D651ABA1ABFD} = {E98514FF-0523-49B4-8564-8D196AE6B640}
 		{7A7C34E6-02DC-448B-BA12-7E259E78FA70} = {1B2FD9A9-A8CE-4AD4-AB94-61A6A49A5788}
 		{C41D0C1D-4400-408D-B45E-82251EF7E4A1} = {E98514FF-0523-49B4-8564-8D196AE6B640}
 		{292874FF-1ECC-4F19-BFB9-B047D26059C5} = {E98514FF-0523-49B4-8564-8D196AE6B640}
+		{3A18D620-D1A6-4188-BF64-48CCF3A8E61D} = {9721A4D2-6BF8-4A73-B2D9-3C416D5A8458}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {71789F54-6517-4F01-AE01-9A05A64BDC5B}
