--- src/Kanban.Application.Dto/Mapper/ClientMapper.cs
@@ -1,16 +1,0 @@
-﻿using App = Kanban.Application.Dto.Models;
-using Repo = Kanban.Repository.Dto.Models;
-
-namespace Kanban.Application.Dto.Mapper;
-
-public static class ClientMapper
-{
-    public static Repo.ClientDto ToDatabase(this App.ClientDto client)
-    {
-        return new Repo.ClientDto
-        {
-            _id = client.Id,
-            Secret = client.Secret,
-        };
-    }
-}
