+++ src/Kanban.Model/Dto/API/Auth/ClientDto.cs
@@ -1,8 +1,7 @@
-﻿namespace Kanban.Application.Dto.Models;
+﻿namespace Kanban.Model.Dto.API.Auth;
 
 public class ClientDto
 {
     public string Id { get; set; } = string.Empty;
-
     public string Secret { get; set; } = string.Empty;
 }
