+++ src/Kanban.Model/Dto/API/Card/CreateCardResponseDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Kanban.Model.Dto.API.Card;
+
+public class CreateCardResponseDto
+{
+    public CardDto? CreatedCard { get; set;}
+}
