+++ src/Kanban.Model/Dto/API/Card/GetCardResponseDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Kanban.Model.Dto.API.Card;
+
+public class GetCardResponseDto
+{
+    public List<CardDto> Cards {  get; set; } = new List<CardDto>();
+}
