+++ src/Kanban.Model/Dto/Application/Card/CardDto.cs
@@ -1,0 +1,10 @@
+﻿namespace Kanban.Model.Dto.Application.Card;
+
+public class CardDto
+{
+    public string Id { get; set; } = string.Empty;
+
+    public string Name { get; set; } = string.Empty;
+
+    public string Description { get; set; } = string.Empty;
+}
