+++ src/Kanban.Model/Dto/Application/Client/ClientDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Kanban.Model.Dto.Application.Client;
+
+public class ClientDto
+{
+    public string Id { get; set; } = string.Empty;
+
+    public string Secret { get; set; } = string.Empty;
+}
