+++ src/Kanban.Model/Dto/Repository/Client/ClientDto.cs
@@ -1,0 +1,10 @@
+﻿using MongoDB.Bson.Serialization.Attributes;
+
+namespace Kanban.Model.Dto.Repository.Client;
+
+public class ClientDto
+{
+    [BsonId]
+    public string _id { get; set; } = string.Empty;
+    public string Secret { get; set; } = string.Empty;
+}
