+++ src/Kanban.Repository/Interfaces/IAuthDatabaseWorker.cs
@@ -1,4 +1,4 @@
-﻿using Kanban.Repository.Dto.Models;
+﻿using Kanban.Model.Dto.Repository.Client;
 
 namespace Kanban.Repository.Interfaces
 {
