+++ src/Kanban.Repository/Kanban.Repository.csproj
@@ -12,7 +12,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Kanban.CrossCutting\Kanban.CrossCutting.csproj" />
-    <ProjectReference Include="..\Kanban.Repository.Dto\Kanban.Repository.Dto.csproj" />
+    <ProjectReference Include="..\Kanban.Model\Kanban.Model.csproj" />
   </ItemGroup>
 
 </Project>
