+++ src/Kanban.Services/Interfaces/ICardService.cs
@@ -1,4 +1,4 @@
-﻿using Kanban.Application.Dto.Models;
+﻿using Kanban.Model.Dto.Application.Card;
 
 namespace Kanban.Application.Interfaces;
