+++ tests/Kanban.Infra.Tests/Usings.cs
@@ -1,7 +1,8 @@
 global using Xunit;
 global using FluentAssertions;
 global using Kanban.Infra.Tests.DatabaseMocks;
-global using Kanban.Repository.Dto.Models;
+global using Kanban.Model.Dto.Repository.Card;
+global using Kanban.Model.Dto.Repository.Client;
 global using MongoDB.Bson;
 global using Newtonsoft.Json;
 using System.Diagnostics.CodeAnalysis;
