+++ tests/Kanban.Presentation.Tests/Kanban.Presentation.Tests.csproj
@@ -25,7 +25,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Kanban.API\Kanban.API.csproj" />
-    <ProjectReference Include="..\..\src\Kanban.Application.Dto\Kanban.Application.Dto.csproj" />
+    <ProjectReference Include="..\..\src\Kanban.Model\Kanban.Model.csproj" />
   </ItemGroup>
 
 </Project>
